Renamed interface
