+++ models/DataObject/Data/Hotspotimage.php
@@ -17,17 +17,17 @@ declare(strict_types=1);
 namespace Pimcore\Model\DataObject\Data;
 
 use Pimcore\Model\Asset;
-use Pimcore\Model\DataObject\InheritanceAwareFieldInterface;
+use Pimcore\Model\DataObject\ReferenceAwareFieldInterface;
 use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
-use Pimcore\Model\DataObject\Traits\InheritanceAwareFieldTrait;
+use Pimcore\Model\DataObject\Traits\ReferenceAwareFieldTrait;
 use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 use Pimcore\Model\Element\ElementDescriptor;
 use Pimcore\Model\Element\Service;
 
-class Hotspotimage implements OwnerAwareFieldInterface, InheritanceAwareFieldInterface
+class Hotspotimage implements OwnerAwareFieldInterface, ReferenceAwareFieldInterface
 {
     use OwnerAwareFieldTrait;
-    use InheritanceAwareFieldTrait;
+    use ReferenceAwareFieldTrait;
 
     protected ElementDescriptor|Asset\Image|null $image = null;
