+++ models/DataObject/Data/ImageGallery.php
@@ -17,16 +17,16 @@ declare(strict_types=1);
 namespace Pimcore\Model\DataObject\Data;
 
 use Iterator;
-use Pimcore\Model\DataObject\InheritanceAwareFieldInterface;
+use Pimcore\Model\DataObject\ReferenceAwareFieldInterface;
 use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
-use Pimcore\Model\DataObject\Traits\InheritanceAwareFieldTrait;
+use Pimcore\Model\DataObject\Traits\ReferenceAwareFieldTrait;
 use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 use Pimcore\Model\Element\Service;
 
-class ImageGallery implements Iterator, OwnerAwareFieldInterface, InheritanceAwareFieldInterface
+class ImageGallery implements Iterator, OwnerAwareFieldInterface, ReferenceAwareFieldInterface
 {
     use OwnerAwareFieldTrait;
-    //use InheritanceAwareFieldTrait;
+    use ReferenceAwareFieldTrait;
 
     /**
      * @var array<int, Hotspotimage|null>
@@ -94,14 +94,5 @@ class ImageGallery implements Iterator, OwnerAwareFieldInterface, InheritanceAwa
         }
 
         return false;
-    }
-
-    public function cloneValue(object $data): object
-    {
-        $context = [
-            'source' => __METHOD__,
-            'conversion' => false,
-        ];
-        return Service::getDeepCopyInstance($data, $context)->copy($data);
     }
 }
