+++ models/DataObject/Data/StructuredTable.php
@@ -17,15 +17,15 @@ declare(strict_types=1);
 namespace Pimcore\Model\DataObject\Data;
 
 use Exception;
-use Pimcore\Model\DataObject\InheritanceAwareFieldInterface;
+use Pimcore\Model\DataObject\ReferenceAwareFieldInterface;
 use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
-use Pimcore\Model\DataObject\Traits\InheritanceAwareFieldTrait;
+use Pimcore\Model\DataObject\Traits\ReferenceAwareFieldTrait;
 use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 
-class StructuredTable implements OwnerAwareFieldInterface, InheritanceAwareFieldInterface
+class StructuredTable implements OwnerAwareFieldInterface, ReferenceAwareFieldInterface
 {
     use OwnerAwareFieldTrait;
-    use InheritanceAwareFieldTrait;
+    use ReferenceAwareFieldTrait;
 
     protected array $data = [];
