+++ models/DataObject/Data/Video.php
@@ -17,19 +17,19 @@ declare(strict_types=1);
 namespace Pimcore\Model\DataObject\Data;
 
 use Pimcore\Model\Asset;
-use Pimcore\Model\DataObject\InheritanceAwareFieldInterface;
+use Pimcore\Model\DataObject\ReferenceAwareFieldInterface;
 use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
-use Pimcore\Model\DataObject\Traits\InheritanceAwareFieldTrait;
+use Pimcore\Model\DataObject\Traits\ReferenceAwareFieldTrait;
 use Pimcore\Model\DataObject\Traits\ObjectVarTrait;
 use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 use Pimcore\Model\Element\ElementDescriptor;
 use Pimcore\Model\Element\Service;
 
-class Video implements OwnerAwareFieldInterface, InheritanceAwareFieldInterface
+class Video implements OwnerAwareFieldInterface, ReferenceAwareFieldInterface
 {
     use OwnerAwareFieldTrait;
     use ObjectVarTrait;
-    use InheritanceAwareFieldTrait;
+    use ReferenceAwareFieldTrait;
 
     protected ?string $type = null;
